requirements.txt
ament-copyright==0.7.12
ament-cppcheck==0.7.12
ament-cpplint==0.7.12
ament-flake8==0.7.12
ament-index-python==0.7.2
ament-lint==0.7.12
ament-lint-cmake==0.7.12
ament-package==0.7.3
ament-pep257==0.7.12
ament-uncrustify==0.7.12
ament-xmllint==0.7.12
anyio==3.6.2
apturl==0.5.2
argcomplete==3.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
async-generator==1.10
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.6.0
bleach==4.1.0
blinker==1.4
Brlapi==0.6.6
catkin-pkg==1.0.0
catkin-pkg-modules==0.5.2
certifi==2018.1.18
cffi==1.15.1
chardet==3.0.4
click==6.7
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.1.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.18.4
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.3.7
contextvars==2.4
coverage==6.2
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
dataclasses==0.8
decorator==4.1.2
defer==1.0.6
defusedxml==0.7.1
demo-nodes-py==0.7.9
distlib==0.3.9
distro==1.9.0
distro-info===0.18ubuntu0.18.04.1
docstring-to-markdown==0.15
docutils==0.14
empy==3.3.2
entrypoints==0.4
examples-rclpy-executors==0.7.5
examples-rclpy-minimal-action-client==0.7.5
examples-rclpy-minimal-action-server==0.7.5
examples-rclpy-minimal-client==0.7.5
examples-rclpy-minimal-publisher==0.7.5
examples-rclpy-minimal-service==0.7.5
examples-rclpy-minimal-subscriber==0.7.5
feedparser==5.2.1
flake8==3.5.0
html5lib==0.999999999
httplib2==0.9.2
idna==3.10
immutables==0.19
importlib-metadata==3.10.1
iniconfig==1.1.1
ipykernel==5.5.6
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.2
jedi-language-server==0.35.1
jetson-stats==4.3.0
Jetson.GPIO==2.0.17
Jinja2==3.0.3
json5==0.9.16
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-lsp==1.5.1
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-lsp==3.9.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
lark-parser==0.7.2
launch==0.8.7
launch-ros==0.8.10
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
lxml==4.2.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
matplotlib==2.1.1
mccabe==0.6.1
mistune==0.8.4
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.6.0
notebook==6.4.10
notify2==0.3.1
numpy==1.19.4
oauth==1.0.1
oauthlib==2.0.6
olefile==0.45.1
onboard==1.4.1
opencv-python==4.6.0.66
openni==2.3.0
osrf-pycommon==0.1.9
packaging==21.3
pandas==0.22.0
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==5.1.0
pkg_resources==0.0.0
pluggy==1.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.36
protobuf==3.0.0
psutil==5.4.2
ptyprocess==0.7.0
py==1.11.0
pycairo==1.16.2
pycodestyle==2.3.1
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
pydantic==1.8.2
pydocstyle==2.0.0
pydot==1.2.3
pyflakes==1.6.0
pygls==0.11.3
Pygments==2.14.0
PyGObject==3.26.1
PyJWT==1.5.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.2.0
pyRFC3339==1.0
pyrsistent==0.18.0
pyserial==3.5
pytest==7.0.1
pytest-cov==4.0.0
pytest-repeat==0.9.1
pytest-rerunfailures==10.3
python-apt==1.6.6
python-dateutil==2.6.1
python-debian==0.1.32
pytz==2018.3
pyxattr==0.6.0
pyxdg==0.25
PyYAML==3.12
pyzmq==25.1.2
quality-of-service-demo-py==0.7.9
requests==2.18.4
requests-unixsocket==0.1.5
roman==2.0.0
ros2action==0.7.11
ros2cli==0.7.11
ros2component==0.7.11
ros2launch==0.8.10
ros2lifecycle==0.7.11
ros2msg==0.7.11
ros2multicast==0.7.11
ros2node==0.7.11
ros2param==0.7.11
ros2pkg==0.7.11
ros2run==0.7.11
ros2service==0.7.11
ros2srv==0.7.11
ros2topic==0.7.11
rosidl-runtime-py==0.7.11
rplidar-roboticia==0.9.5
rqt-action==0.4.9
rqt-graph==1.0.4
rqt-gui==1.0.7
rqt-gui-py==1.0.7
rqt-reconfigure==1.0.7
rqt-topic==0.4.9
scipy==0.19.1
SecretStorage==2.3.1
Send2Trash==1.8.3
simplejson==3.13.2
six==1.11.0
smbus2==0.5.0
sniffio==1.2.0
snowballstemmer==1.2.1
sros2==0.7.2
ssh-import-id==5.7
system-service==0.3
systemd-python==234
teleop-twist-keyboard==2.3.2
terminado==0.12.1
testpath==0.6.0
tomli==1.2.3
topic-monitor==0.7.9
torch @ file:///home/tx2/jetson_tx2_installation/torch-1.10.0-cp36-cp36m-linux_aarch64.whl
tornado==6.1
traitlets==4.3.3
typeguard==2.13.3
typing_extensions==4.1.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unity-scope-calculator==0.1
unity-scope-chromiumbookmarks==0.1
unity-scope-colourlovers==0.1
unity-scope-devhelp==0.1
unity-scope-firefoxbookmarks==0.1
unity-scope-manpages==0.1
unity-scope-openclipart==0.1
unity-scope-texdoc==0.1
unity-scope-tomboy==0.1
unity-scope-virtualbox==0.1
unity-scope-yelp==0.1
unity-scope-zotero==0.1
urllib3==1.22
urwid==2.0.1
virtualenv==15.1.0
wadllib==1.3.2
wcwidth==0.2.13
webencodings==0.5
websocket-client==1.3.1
xkit==0.0.0
youtube_dl==2018.3.14
zipp==3.6.0
zope.interface==4.3.2
